Split off EntityTweener into seperate file
