+++ src/openrct2/ReplayManager.cpp
@@ -29,6 +29,7 @@
 #include "object/ObjectManager.h"
 #include "object/ObjectRepository.h"
 #include "rct2/S6Exporter.h"
+#include "world/EntityTweener.h"
 #include "world/Park.h"
 #include "world/Sprite.h"
 #include "zlib.h"
