+++ src/openrct2/peep/Peep.cpp
@@ -36,6 +36,7 @@
 #include "../windows/Intent.h"
 #include "../world/Balloon.h"
 #include "../world/Climate.h"
+#include "../world/EntityTweener.h"
 #include "../world/Entrance.h"
 #include "../world/Footpath.h"
 #include "../world/LargeScenery.h"
